Refactor demo into src. Add descriptive analysis
