--- demo/models/metadata.json
@@ -1,20 +1,0 @@
-{
-  "CLIENTE1": {
-    "model_path": "models/CLIENTE1.keras",
-    "features": [
-      "Volumen",
-      "Temperatura",
-      "Presion"
-    ],
-    "lookback": 24
-  },
-  "CLIENTE2": {
-    "model_path": "models/CLIENTE2.keras",
-    "features": [
-      "Volumen",
-      "Temperatura",
-      "Presion"
-    ],
-    "lookback": 24
-  }
-}
