--- demo/requirements.txt
@@ -1,8 +1,0 @@
-tensorflow>=2.13
-pandas>=2.0
-numpy>=1.24
-scikit-learn>=1.3
-streamlit>=1.31
-plotly>=5.20
-openpyxl>=3.1
-requests>=2.31
